Move useUsersListener into user/hooks
