+++ twake/frontend/src/app/services/user/hooks/useUser.ts
@@ -1,0 +1,23 @@
+import { useCallback, useEffect, useState } from 'react';
+import { UserType } from 'app/models/User';
+import userGetAsync from '../AsyncGet';
+
+const useUser = (userId: string): { user: UserType | undefined } => {
+  const [user, setUser] = useState<UserType>();
+
+  const fetchUser = useCallback(async () => {
+    const user = await userGetAsync(userId);
+
+    user && setUser(user);
+  }, [userId]);
+
+  useEffect(() => {
+    fetchUser();
+  }, [fetchUser]);
+
+  return {
+    user,
+  };
+};
+
+export default useUser;
